Do auth check before running all page queries.
Doh!  This is much more efficient than running the queries and
throwing away the results if the user has not passed the auth
check.  Ugh.  At least I could do this in one place and not rework
the logic in the routes which was my first solution to this problem.
